--- src/models/entity-types.ts
@@ -1,2 +1,0 @@
-
-export const entityTypes: string [] = ['donor', 'specimen', 'sample', 'treatment', 'primary_diagnosis', 'chemotherapy', 'hormone_therapy'];
