+++ src/server.ts
@@ -11,10 +11,9 @@ const port = process.env.PORT
 
 import * as express from 'express';
 import {getIdForEntity, root} from "./routes/root";
-//import {AppDataSource} from "./data-source";
 import {defaultErrorHandler} from "./middlewares/error-handler";
 import {getSchemaInfo} from "./services/id-service";
-import {getConnectionAndRepo} from "./middlewares/repo-connection";
+import {getConnectionAndRepo} from "./middlewares/datasource";
 
 
 const cors = require("cors");
@@ -26,8 +25,8 @@ function setupExpress(){
     app.use(bodyParser.json());
 
     app.route("/").get(root);
-    app.route("/:programId/:submitterId/:entityType").get(getIdForEntity);
-
+    //app.route("/:programId/:submitterId/:entityType").get(getIdForEntity);
+    app.route(process.env["REQUEST_ROUTE"]).get(getIdForEntity);
     app.use(defaultErrorHandler);
 
 }
@@ -40,7 +39,6 @@ function startServer(){
 
 function initializeDB(){
     const entities: [] = JSON.parse(process.env["ENTITY_LIST"]);
-
     entities.forEach((entity) => {
         const  schemaInfo = getSchemaInfo(entity);
         const repo = getConnectionAndRepo(schemaInfo, 1)
@@ -52,48 +50,5 @@ function initializeDB(){
 setupExpress();
 startServer();
 initializeDB();
-
-// --------------------------------------------
-/*
-
-
-export interface data {
-    [key: string]: string;
-}
-
-let schemaInfo = {} as SchemaInfo;
-let schemaInfo2 = {} as SchemaInfo;
-
-
-
-const columns1: {name: string, type: ColumnType}[] = JSON.parse(process.env.DONOR_TABLE_COLUMNS)
-const columns2: {name: string, type: ColumnType}[] = JSON.parse(process.env.SAMPLE_TABLE_COLUMNS)
-
-
-schemaInfo.tablename =  process.env.DONOR_TABLE_NAME; //"TestTable1";
-schemaInfo.columns = columns1;
-
-
-schemaInfo2.tablename = process.env.SAMPLE_TABLE_NAME; ///"TestTable2";
-const x = process.env[`SAMPLE_TABLE_NAME` ];
-schemaInfo2.columns = columns2;
-
-async function createData(repo: Repository<ObjectLiteral>, data: any) {
-    await repo.save(data);
-    //await destroyConnection();
-
-}
-
-
-const data1: data = {column1: "table1Value"};
-getConnectionAndRepo(schemaInfo).then(repo => {createData(repo, data1)}) ;
-
-const data2: data  = {column1: "table2Value1", column3: "table2Value3"}
-getConnectionAndRepo(schemaInfo2).then(repo => {createData(repo, data2)}) ;
-
-*/
-
-
-// --------------------------------------------
