--- src/services/dynamic-id-service.ts
@@ -1,15 +1,0 @@
-import {ColumnType} from "typeorm";
-//import {createTableAndRepo} from "../models/dynamic-table";
-
-/*function createSchema(){
-
-    const columns: {name: string, type: ColumnType}[] = [{
-        "name": "column1",
-        "type": "string"
-    }]
-
-    createTableAndRepo("TestTable", columns);
-}*/
-
-
-
